Implement client, server, some tests

+++ hw/ftp/src/main/java/ru/spbau/bocharov/ftp/protocol/MessageType.java
@@ -1,6 +1,6 @@
 package ru.spbau.bocharov.ftp.protocol;
 
 public class MessageType {
-    public static final int LIST = 0;
-    public static final int GET = 1;
+    public static final byte LIST = 0;
+    public static final byte GET = 1;
 }
